Space Invaders Python/requirements.txt
psycopg2-binary==2.9.9
pygame==2.6.0
